refactor : (다리 건너기) 인풋 클래스, 아웃풋 클래스 view 패키지로 이동
