+++ src/main/java/bridge/view/InputView.java
@@ -1,4 +1,4 @@
-package bridge.input;
+package bridge.view;
 
 import camp.nextstep.edu.missionutils.Console;
