+++ src/main/java/bridge/view/OutputView.java
@@ -1,7 +1,8 @@
-package bridge.output;
+package bridge.view;
 
 
 import bridge.constants.Command;
+import bridge.domain.BridgeState;
 import java.util.HashMap;
 
 /**
@@ -14,9 +15,8 @@ public class OutputView {
      * <p>
      * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    public void printMap(HashMap<String, StringBuilder> bridgeConnection) {
-        System.out.println(bridgeConnection.get(Command.UP.getCommand()));
-        System.out.println(bridgeConnection.get(Command.DOWN.getCommand()));
+    public void printMap(BridgeState bridgePlace) {
+        bridgePlace.printBridge();
     }
 
 
@@ -25,13 +25,15 @@ public class OutputView {
      * <p>
      * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    public void printResult(HashMap<String, StringBuilder> bridgeState, int count, boolean success) {
-        System.out.println("최종 게임 결과");
-        System.out.println(bridgeState.get(Command.UP.getCommand()));
-        System.out.println(bridgeState.get(Command.DOWN.getCommand()));
-
+    public void printResult(BridgeState bridgeState, int count, boolean success) {
+        bridgeState.endPrintBridge();
         System.out.println();
 
+        gameSuccess(success);
+        System.out.println("총 시도한 횟수: " + count);
+    }
+
+    private void gameSuccess(boolean success) {
         if (success == true) {
             System.out.println("게임 성공 여부: 성공");
         }
@@ -39,8 +41,6 @@ public class OutputView {
         if (success == false) {
             System.out.println("게임 성공 여부: 실패");
         }
-
-        System.out.println("총 시도한 횟수: " + count);
     }
 }
